# openapi/route-service.yaml
openapi: 3.0.3
info:
  title: Route Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080
x-curl-examples:
  listCities:
    description: List seeded cities.
    command: |
      curl http://localhost:8080/api/cities
  getCity:
    description: Get seeded Ekaterinburg city by id.
    command: |
      curl http://localhost:8080/api/cities/11111111-1111-1111-1111-111111111111
  listRoutes:
    description: List seeded routes.
    command: |
      curl http://localhost:8080/api/routes
  getRoute:
    description: Get seeded route by id.
    command: |
      curl http://localhost:8080/api/routes/33333333-3333-3333-3333-333333333333
  searchRoutes:
    description: Search route from Ekaterinburg to Chelyabinsk.
    command: |
      curl "http://localhost:8080/api/routes/search?fromCityId=11111111-1111-1111-1111-111111111111&toCityId=22222222-2222-2222-2222-222222222222"
  createCityAsAdmin:
    description: Create city, requires ADMIN token.
    command: |
      curl -X POST http://localhost:8080/api/cities \
        -H "Content-Type: application/json" \
        -H "Authorization: Bearer $ADMIN_TOKEN" \
        -d '{"name":"Tyumen","region":"Tyumen Oblast","country":"Russia","active":true}'
paths:
  /api/cities:
    get:
      summary: List cities
      operationId: listCities
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
    post:
      summary: Create city
      operationId: createCity
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/cities/{id}:
    get:
      summary: Get city
      operationId: getCityById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update city
      operationId: updateCity
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete city
      operationId: deleteCity
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/routes:
    get:
      summary: List routes
      operationId: listRoutes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
    post:
      summary: Create route
      operationId: createRoute
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete route
      operationId: deleteRoute
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/routes/{id}:
    get:
      summary: Get route
      operationId: getRouteById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update route
      operationId: updateRoute
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/routes/search:
    get:
      summary: Search routes
      operationId: searchRoutes
      parameters:
        - name: fromCityId
          in: query
          schema: { type: string, format: uuid }
        - name: toCityId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CityRequest:
      type: object
      properties:
        name:
          type: string
          example: Tyumen
        region:
          type: string
          example: Tyumen Oblast
        country:
          type: string
          example: Russia
        active:
          type: boolean
          default: true
    City:
      allOf:
        - $ref: '#/components/schemas/CityRequest'
        - type: object
          required: [id, active]
          properties:
            id:
              type: string
              format: uuid
    RouteRequest:
      type: object
      properties:
        fromCityId:
          type: string
          format: uuid
        toCityId:
          type: string
          format: uuid
        distanceKm:
          type: integer
          format: int32
          minimum: 0
        estimatedDurationMinutes:
          type: integer
          format: int32
          minimum: 0
        active:
          type: boolean
          default: true
    Route:
      allOf:
        - $ref: '#/components/schemas/RouteRequest'
        - type: object
          required: [id, active]
          properties:
            id:
              type: string
              format: uuid
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string